Moving all hooks test under the hooks folder
