+++ karate-core/src/test/java/com/intuit/karate/core/hooks/ScenarioHookTest.java
@@ -1,8 +1,9 @@
-package com.intuit.karate.core.runner;
+package com.intuit.karate.core.hooks;
 
 import com.intuit.karate.Results;
 import com.intuit.karate.Runner;
 import static org.junit.jupiter.api.Assertions.*;
+
 import org.junit.jupiter.api.Test;
 
 /**
@@ -13,7 +14,7 @@ class ScenarioHookTest {
 
     @Test
     void testStopIfScenarioHasNoTags() {
-        String path = "classpath:com/intuit/karate/core/runner/test-hook-notags.feature";
+        String path = "classpath:com/intuit/karate/core/hooks/test-hook-notags.feature";
         Results results = Runner.path(path).hook(new MandatoryTagHook()).parallel(1);
         assertEquals(1, results.getFeaturesTotal());
         assertEquals(1, results.getFailCount());
@@ -21,7 +22,7 @@ class ScenarioHookTest {
 
     @Test
     void testHookForExamplesWithTags() {
-        String path = "classpath:com/intuit/karate/core/runner/test-hook-multiexample.feature";
+        String path = "classpath:com/intuit/karate/core/hooks/test-hook-multiexample.feature";
         Results results = Runner.path(path).hook(new MandatoryTagHook()).parallel(1);
         assertEquals(1, results.getFeaturesTotal());
         assertEquals(7, results.getScenariosTotal());
